Removing Bitcoin Support
We're removing Bitcoin support for exclusive Monero-only support going forward. We originally introduced Bitcoin as an alternative support method for those requesting it, but it received almost no usage, Monero saw over 30x the number of transactions; all with the benefit of better privacy for our audience. If using Monero is an issue, we suggest users refer to https://kycnot.me/ for methods of converting their cryptocurrency to Monero. Thanks for working with us